+++ src/main/c/cext/call.c
@@ -190,6 +190,10 @@ int rb_obj_method_arity(VALUE object, ID id) {
   return polyglot_as_i32(RUBY_CEXT_INVOKE_NO_WRAP("rb_obj_method_arity", object, ID2SYM(id)));
 }
 
+VALUE rb_obj_method(VALUE object, VALUE id) {
+  return RUBY_CEXT_INVOKE("rb_obj_method", object, id);
+}
+
 int rb_obj_respond_to(VALUE object, ID id, int priv) {
   return polyglot_as_boolean(polyglot_invoke(RUBY_CEXT, "rb_obj_respond_to", rb_tr_unwrap(object), rb_tr_id2sym(id), priv));
 }