+++ src/main/java/org/truffleruby/core/module/ModuleNodes.java
@@ -1846,6 +1846,7 @@ public abstract class ModuleNodes {
     @NodeChild(value = "module", type = RubyNode.class)
     @NodeChild(value = "name", type = RubyNode.class)
     public abstract static class InstanceMethodNode extends CoreMethodNode {
+        @Child private ReadCallerFrameNode readCallerFrame = ReadCallerFrameNode.create();
 
         @CreateCast("name")
         protected RubyNode coerceToString(RubyNode name) {
@@ -1853,10 +1854,13 @@ public abstract class ModuleNodes {
         }
 
         @Specialization
-        protected RubyUnboundMethod instanceMethod(RubyModule module, String name,
+        protected RubyUnboundMethod instanceMethod(VirtualFrame frame, RubyModule module, String name,
                 @Cached BranchProfile errorProfile) {
+            final Frame callerFrame = readCallerFrame.execute(frame);
+            final DeclarationContext declarationContext = RubyArguments.getDeclarationContext(callerFrame);
+
             // TODO(CS, 11-Jan-15) cache this lookup
-            final InternalMethod method = ModuleOperations.lookupMethodUncached(module, name, null);
+            final InternalMethod method = ModuleOperations.lookupMethodUncached(module, name, declarationContext);
 
             if (method == null || method.isUndefined()) {
                 errorProfile.enter();